move global banner values
