+++ shared/global_banner_settings.js
